+++ extensions/subwatch/src/add-subscription.tsx
@@ -1,23 +1,30 @@
-import { Action, ActionPanel, Form, getPreferenceValues, showToast, Toast, useNavigation } from "@raycast/api";
-import { getStuff } from "./state";
+import {
+  Action,
+  ActionPanel,
+  Form,
+  getPreferenceValues,
+  showToast,
+  Toast,
+  useNavigation,
+} from "@raycast/api";
+import { fetchSubscriptions } from "./state";
 import { NewSubscription } from "./types";
-import { FormValidation, useForm } from "@raycast/utils";
+import { FormValidation, showFailureToast, useForm } from "@raycast/utils";
 
 export default function Command() {
   const { pop } = useNavigation();
-  const { subwatchApiKey } = getPreferenceValues<Preferences>();
-  const { isLoading, data, error, mutate, revalidate } = getStuff();
+  const { subwatchApiKey, supabaseApiKey } = getPreferenceValues<Preferences>();
+  const { isLoading, data, error, mutate, revalidate } = fetchSubscriptions();
 
   if (error) {
-    showToast({
-      style: Toast.Style.Success,
-      title: "An error occurred!",
-      message: error.message,
-    });
+    showFailureToast(error, { title: "An error occurred!" });
   }
 
   async function handleCreate(subscription: NewSubscription) {
-    const toast = await showToast({ style: Toast.Style.Animated, title: `Adding ${subscription.service}` });
+    await showToast({
+      style: Toast.Style.Animated,
+      title: `Adding ${subscription.service}`,
+    });
     try {
       data?.[0].data.push({
         name: subscription.service,
@@ -25,8 +32,8 @@ export default function Command() {
         count: 0,
         billing: [
           {
-            start_date: String(subscription.startDate.toISOString()),
-            end_date: String(subscription.endDate),
+            start_date: String(subscription.start_date.toISOString()),
+            end_date: String(subscription.end_date),
             price: Number(subscription.price),
             interval: subscription.interval,
             count: 0,
@@ -35,15 +42,20 @@ export default function Command() {
       });
 
       await mutate(
-        fetch("https://nzyzephaenhlxoohrphc.supabase.co/rest/v1/rpc/raycast_update_data", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            apikey:
-              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im56eXplcGhhZW5obHhvb2hycGhjIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDQxOTQzMjgsImV4cCI6MjA1OTc3MDMyOH0.6AboCGgJGqJMTgqUH3LsYmhoWQ8sfEWqdv0cY-1EXIg",
+        fetch(
+          "https://nzyzephaenhlxoohrphc.supabase.co/rest/v1/rpc/raycast_update_data",
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              apikey: supabaseApiKey,
+            },
+            body: JSON.stringify({
+              raycast_uuid: subwatchApiKey,
+              newdata: data?.[0].data,
+            }),
           },
-          body: JSON.stringify({ raycast_uuid: subwatchApiKey, newdata: data?.[0].data }),
-        }),
+        ),
       );
       revalidate();
 
@@ -54,15 +66,15 @@ export default function Command() {
       });
     } catch (err) {
       // The data will automatically be rolled back to its previous value.
-      toast.style = Toast.Style.Failure;
-      toast.title = `Could not add ${subscription.service}`;
-      toast.message = err?.message;
+      showFailureToast(err, { title: `Could not add ${subscription.service}` });
     }
   }
 
   const { handleSubmit, itemProps } = useForm<NewSubscription>({
     async onSubmit(values) {
-      const cleanedDomain = values.domain.replace(/https?:\/\//gm, "").replace(/www\./gm, "");
+      const cleanedDomain = values.domain
+        .replace(/https?:\/\//gm, "")
+        .replace(/www\./gm, "");
       handleCreate({ ...values, domain: cleanedDomain });
       pop();
     },
@@ -74,10 +86,12 @@ export default function Command() {
           return "Price should be a number";
         } else if (!value) {
           return "The item is required";
+        } else if (Number(value) < 1) {
+          return "Price should be a positive number";
         }
       },
       interval: FormValidation.Required,
-      startDate: FormValidation.Required,
+      start_date: FormValidation.Required,
     },
   });
 
@@ -87,14 +101,25 @@ export default function Command() {
       actions={
         <ActionPanel>
           <ActionPanel.Section title="Copy">
-            <Action.SubmitForm title="Create New Subscription" onSubmit={handleSubmit} />
+            <Action.SubmitForm
+              title="Create New Subscription"
+              onSubmit={handleSubmit}
+            />
           </ActionPanel.Section>
         </ActionPanel>
       }
     >
       <Form.Description text="Track, alert, and optimize a new subscription" />
-      <Form.TextField title="Service" placeholder="apple" {...itemProps.service} />
-      <Form.TextField title="Domain" placeholder="apple.com" {...itemProps.domain} />
+      <Form.TextField
+        title="Service"
+        placeholder="apple"
+        {...itemProps.service}
+      />
+      <Form.TextField
+        title="Domain"
+        placeholder="apple.com"
+        {...itemProps.domain}
+      />
 
       <Form.TextField title="Price" placeholder="$8" {...itemProps.price} />
 
@@ -103,8 +128,8 @@ export default function Command() {
         <Form.Dropdown.Item value="yearly" title="Yearly" />
       </Form.Dropdown>
 
-      <Form.DatePicker id="startDate" title="Start Date" />
-      <Form.DatePicker id="endDate" title="End Date" />
+      <Form.DatePicker id="start_date" title="Start Date" />
+      <Form.DatePicker id="end_date" title="End Date" />
     </Form>
   );
 }