+++ extensions/subwatch/src/manage-subscriptions.tsx
@@ -1,36 +1,49 @@
-import { Action, ActionPanel, getPreferenceValues, Icon, List, showToast, Toast } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  getPreferenceValues,
+  Icon,
+  List,
+  showToast,
+  Toast,
+} from "@raycast/api";
 
-import { getStuff } from "./state";
+import { fetchSubscriptions } from "./state";
 import CreateSubscriptionAction from "./add-subscription";
+import { showFailureToast } from "@raycast/utils";
 
 export default function Command() {
-  const { subwatchApiKey } = getPreferenceValues<Preferences>();
-  const { isLoading, data, error, mutate } = getStuff();
+  const { subwatchApiKey, supabaseApiKey } = getPreferenceValues<Preferences>();
+  const { isLoading, data, error, mutate } = fetchSubscriptions();
 
   if (error) {
-    showToast({
-      style: Toast.Style.Success,
-      title: "An error occurred!",
-      message: error.message,
-    });
+    showFailureToast(error, { title: "An error occurred!" });
   }
 
   async function handleDelete(index: number) {
     const subscriptionToDelete = data?.[0].data[index];
-    const toast = await showToast({ style: Toast.Style.Animated, title: `Deleting ${subscriptionToDelete?.name}` });
+    await showToast({
+      style: Toast.Style.Animated,
+      title: `Deleting ${subscriptionToDelete?.name}`,
+    });
 
     try {
       data?.[0].data.splice(index, 1);
       await mutate(
-        fetch("https://nzyzephaenhlxoohrphc.supabase.co/rest/v1/rpc/raycast_update_data", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            apikey:
-              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im56eXplcGhhZW5obHhvb2hycGhjIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDQxOTQzMjgsImV4cCI6MjA1OTc3MDMyOH0.6AboCGgJGqJMTgqUH3LsYmhoWQ8sfEWqdv0cY-1EXIg",
+        fetch(
+          "https://nzyzephaenhlxoohrphc.supabase.co/rest/v1/rpc/raycast_update_data",
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              apikey: supabaseApiKey,
+            },
+            body: JSON.stringify({
+              raycast_uuid: subwatchApiKey,
+              newdata: data?.[0].data,
+            }),
           },
-          body: JSON.stringify({ raycast_uuid: subwatchApiKey, newdata: data?.[0].data }),
-        }),
+        ),
       );
       showToast({
         style: Toast.Style.Success,
@@ -39,9 +52,9 @@ export default function Command() {
       });
     } catch (err) {
       // The data will automatically be rolled back to its previous value.
-      toast.style = Toast.Style.Failure;
-      toast.title = `Could not add ${subscriptionToDelete?.name}`;
-      toast.message = err?.message;
+      showFailureToast(err, {
+        title: `Could not delete ${subscriptionToDelete?.name}`,
+      });
     }
   }
 
@@ -72,7 +85,9 @@ export default function Command() {
                   shortcut={{ modifiers: ["cmd"], key: "n" }}
                   target={<CreateSubscriptionAction />}
                 />
-                <DeleteSubscriptionAction onDelete={() => handleDelete(index)} />
+                <DeleteSubscriptionAction
+                  onDelete={() => handleDelete(index)}
+                />
               </ActionPanel.Section>
             </ActionPanel>
           }
@@ -81,15 +96,31 @@ export default function Command() {
               markdown={`![Logo](https://img.logo.dev/${item.domain || `${item.name}.com`}?token=pk_JrIah0kcTFeKu4Xk9or1xw)`}
               metadata={
                 <List.Item.Detail.Metadata>
-                  <List.Item.Detail.Metadata.Link title="Domain" target={`https://${item.domain}`} text={item.domain} />
-                  <List.Item.Detail.Metadata.Label title="Interval" text={item.billing[0].interval} />
-                  <List.Item.Detail.Metadata.Label title="Pricing" text={String(item.billing[0].price)} />
+                  <List.Item.Detail.Metadata.Link
+                    title="Domain"
+                    target={`https://${item.domain}`}
+                    text={item.domain}
+                  />
+                  <List.Item.Detail.Metadata.Label
+                    title="Interval"
+                    text={item.billing[0].interval}
+                  />
+                  <List.Item.Detail.Metadata.Label
+                    title="Pricing"
+                    text={String(item.billing[0].price)}
+                  />
                   <List.Item.Detail.Metadata.Separator />
-                  <List.Item.Detail.Metadata.Label title="Start date" text={item.billing[0].start_date} />
+                  <List.Item.Detail.Metadata.Label
+                    title="Start date"
+                    text={item.billing[0].start_date}
+                  />
                   <List.Item.Detail.Metadata.Label
                     title="End date"
                     text={
-                      item.billing[0].end_date && item.billing[0].end_date != "null" ? item.billing[0].end_date : ""
+                      item.billing[0].end_date &&
+                      item.billing[0].end_date != "null"
+                        ? item.billing[0].end_date
+                        : ""
                     }
                   />
                 </List.Item.Detail.Metadata>